+++ Editor/InputPropertyDrawer.cs
@@ -1,7 +1,7 @@
 ﻿// NOTE: Put in an editor folder. //
 
 // Originally created by Denis Rizov
-// Modified by kiltec
+// Modified by Kiltec
 
 using UnityEngine;
 using UnityEditor;