+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ---
 
-0.0.2 Updated for Unity Packages
+0.0.13 Added MinMaxRange Attribute
 
 ---
 
-0.0.1 Initial version
+0.0.12 Updated for Unity Packages
+
+---
+
+0.0.11 Initial version extracted from UnityTools